+++ syndicate/connection/s3_connection.py
@@ -50,10 +50,8 @@ class S3Connection(object):
     def download_file(self, bucket_name, key, file_path):
         self.resource.Bucket(bucket_name).download_file(key, file_path)
 
-    def download_to_file(self, bucket_name, key):
-        buffer = io.BytesIO()
-        self.resource.Bucket(bucket_name).download_fileobj(key, buffer)
-        return buffer
+    def download_to_file(self, bucket_name, key, file):
+        self.resource.Bucket(bucket_name).download_fileobj(key, file)
 
     def upload_file(self, storage, file_name, bucket_name, folder=''):
         """ Upload specific file to s3.
@@ -67,15 +65,17 @@ class S3Connection(object):
         self.client.upload_file(storage + file_name, bucket_name,
                                 folder + file_name)
 
-    def upload_single_file(self, path, key, bucket):
+    def upload_single_file(self, path, key, bucket, extra_args=None):
         """ Uploads file just like method above, but allows to specify
         object key as argument
 
         :param path: path to actual file on os
         :param key: object key, as it will be uploaded to bucket
         :param bucket: just bucket name
+        :param extra_args: Extra arguments that may be passed to the client
+        operation
         """
-        self.client.upload_file(path, bucket, key)
+        self.client.upload_file(path, bucket, key, ExtraArgs=extra_args)
 
     def put_object(self, file_obj, key, bucket,
                    content_type, content_encoding=None):
@@ -122,6 +122,15 @@ class S3Connection(object):
             return self.client.get_bucket_policy(Bucket=bucket_name)
         except ClientError as e:
             if 'NoSuchBucketPolicy' or 'NoSuchBucket' in str(e):
+                pass  # valid exception
+            else:
+                raise e
+
+    def get_bucket_website(self, bucket_name):
+        try:
+            return self.client.get_bucket_website(Bucket=bucket_name)
+        except ClientError as e:
+            if 'NoSuchWebsiteConfiguration' or 'NoSuchBucket' in str(e):
                 pass  # valid exception
             else:
                 raise e