+++ syndicate/core/resources/s3_resource.py
@@ -126,6 +126,7 @@ class S3Resource(BaseResource):
 
         policy = meta.get('policy')
         if policy:
+            self._populate_bucket_name_in_policy(policy, name)
             self.s3_conn.add_bucket_policy(name, policy)
             _LOG.debug('Policy on {0} S3 bucket is set up.'.format(name))
 
@@ -144,8 +145,9 @@ class S3Resource(BaseResource):
             _LOG.debug(f'Website hosting configured with parameters: '
                        f'\'index_document\': \'{index_document}\', '
                        f'\'error_document\': \'{error_document}\'')
-            website_endpoint = (f'http://{name}.s3-website.'
-                                f'{self.s3_conn.region}.amazonaws.com')
+            website_endpoint = (
+                f'http://{name}.s3-website.{self.s3_conn.region}.amazonaws.com'
+                f'/{index_document}')
             USER_LOG.info(f'Bucket website endpoint: {website_endpoint}')
 
         rules = meta.get('LifecycleConfiguration')
@@ -214,3 +216,18 @@ class S3Resource(BaseResource):
         # TODO add files keys support to regex
         return bool(re.match(r'^arn:aws:s3:::[a-z0-9.-]{3,63}/?$',
                              maybe_arn))
+
+    @staticmethod
+    def _populate_bucket_name_in_policy(policy, bucket_name):
+        statements = policy['Statement'] if policy.get('Statement') else []
+        for statement in statements:
+            resources = statement['Resource'] if statement.get('Resource') \
+                else []
+            new_resources = []
+            for resource in resources:
+                if 'bucket_name' in resource:
+                    new_resources.append(resource.format(bucket_name=
+                                                         bucket_name))
+                else:
+                    new_resources.append(resource)
+            statement['Resource'] = new_resources