+++ syndicate/core/resources/swagger_ui_resource.py
@@ -13,6 +13,7 @@
     See the License for the specific language governing permissions and
     limitations under the License.
 """
+import io
 import os
 from pathlib import PurePath
 from shutil import rmtree
@@ -21,7 +22,7 @@ from zipfile import ZipFile
 from syndicate.commons import deep_get
 from syndicate.commons.log_helper import get_logger, get_user_logger
 from syndicate.core.constants import SWAGGER_UI_ARTIFACT_NAME_TEMPLATE, \
-    ARTIFACTS_FOLDER
+    ARTIFACTS_FOLDER, SWAGGER_UI_SPEC_NAME_TEMPLATE
 from syndicate.core.helper import build_path, unpack_kwargs
 from syndicate.core.resources.base_resource import BaseResource
 from syndicate.core.resources.helper import build_description_obj
@@ -35,12 +36,15 @@ USER_LOG = get_user_logger()
 class SwaggerUIResource(BaseResource):
 
     def __init__(self, s3_conn, deploy_target_bucket, region,
-                 account_id) -> None:
+                 account_id, extended_prefix_mode, prefix, suffix) -> None:
         from syndicate.core import CONF_PATH
         self.s3_conn = s3_conn
         self.deploy_target_bucket = deploy_target_bucket
         self.region = region
         self.account_id = account_id
+        self.extended_prefix_mode = extended_prefix_mode
+        self.prefix = prefix
+        self.suffix = suffix
         self.conf_path = CONF_PATH
 
     def create_update_swagger_ui(self, args):
@@ -50,9 +54,15 @@ class SwaggerUIResource(BaseResource):
         return self.create_pool(self._remove_swagger_ui, args)
 
     @unpack_kwargs
-    def _create_update_swagger_ui_from_meta(self, name, meta):
+    def _create_update_swagger_ui_from_meta(self, name, meta, context=None):
+        pure_name = name
+        if self.extended_prefix_mode:
+            if self.prefix:
+                pure_name = name[len(self.prefix):]
+            if self.suffix:
+                pure_name = pure_name[:-len(self.suffix)]
         artifact_file_name = SWAGGER_UI_ARTIFACT_NAME_TEMPLATE.format(
-            name=name)
+            name=pure_name)
         artifact_dir = PurePath(self.conf_path, ARTIFACTS_FOLDER,
                                 name).as_posix()
         target_bucket = meta.get('target_bucket')
@@ -74,28 +84,65 @@ class SwaggerUIResource(BaseResource):
             artifact_src_path = bundle_path + '/' + artifact_file_name
 
         _LOG.info(f'Downloading an artifact for Swagger UI \'{name}\'')
-        artifact = self.s3_conn.download_to_file(
-            bucket_name=deploy_bucket_name,
-            key=artifact_src_path)
-
-        extract_to = build_path(artifact_dir, name)
-        with ZipFile(artifact, 'r') as zf:
-            zf.extractall(extract_to)
+        with io.BytesIO() as artifact:
+            self.s3_conn.download_to_file(
+                    bucket_name=deploy_bucket_name,
+                    key=artifact_src_path,
+                    file=artifact)
+            extract_to = build_path(artifact_dir, name)
+            with ZipFile(artifact, 'r') as zf:
+                zf.extractall(extract_to)
 
         _LOG.info(f'Uploading files for Swagger UI \'{name}\' to target '
                   f'bucket \'{target_bucket}\'')
         for file in os.listdir(extract_to):
+            extra_args = None
+            if file == INDEX_FILE_NAME:
+                extra_args = {
+                    'ContentType': 'text/html',
+                    'ContentDisposition': f'inline;filename={INDEX_FILE_NAME}'
+                }
             self.s3_conn.upload_single_file(path=PurePath(extract_to,
                                                           file).as_posix(),
                                             key=file,
-                                            bucket=target_bucket)
+                                            bucket=target_bucket,
+                                            extra_args=extra_args)
 
         _LOG.info(f'Removing temporary directory \'{artifact_dir}\'')
         rmtree(artifact_dir)
 
+        return self.describe_swagger_ui(name, meta)
+
+    def describe_swagger_ui(self, name, meta):
+        target_bucket = meta.get('target_bucket')
+        arn = (f'arn:aws-syndicate:{self.region}:{self.account_id}:'
+               f'{target_bucket}')
+        spec_file_name = SWAGGER_UI_SPEC_NAME_TEMPLATE.format(
+            name=name)
+        website_hosting = self.s3_conn.get_bucket_website(target_bucket)
+        bucket_description = {
+            'arn': f'arn:aws:s3:::{target_bucket}',
+            'bucket_acl': self.s3_conn.get_bucket_acl(target_bucket),
+            'location': self.s3_conn.get_bucket_location(target_bucket),
+            'policy': self.s3_conn.get_bucket_policy(target_bucket)
+        }
+        response = {
+            'host_description': bucket_description
+        }
+        if website_hosting:
+            hosting_config = {
+                "enabled": True,
+                "index_document": deep_get(website_hosting,
+                                           ['IndexDocument', 'Suffix'])
+            }
+            if self.s3_conn.is_file_exists(target_bucket, spec_file_name):
+                hosting_config['api_spec_document'] = spec_file_name
+            hosting_config['endpoint'] = (
+                f'http://{target_bucket}.s3-website.{self.region}.'
+                f'amazonaws.com')
+            response['website_hosting'] = hosting_config
         return {
-            f'arn:aws-syndicate:{self.region}:{self.account_id}:{name}':
-                build_description_obj(None, name, meta)
+            arn: build_description_obj(response, name, meta)
         }
 
     @unpack_kwargs